move the commands back into the CLI dir, i don't think they need their own dir
